requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
tk
